Removed some duplicate css
